requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
supabase==2.28.2
apscheduler==3.10.4
python-telegram-bot==21.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
feedparser==6.0.11
